+++ mayan/__init__.py
@@ -1,9 +1,9 @@
 from __future__ import unicode_literals
 
 __title__ = 'Mayan EDMS'
-__version__ = '3.1.1'
+__version__ = '3.1.2'
 __build__ = 0x030101
-__build_string__ = 'v3.1.1_Tue Sep 18 22:10:17 2018 -0400'
+__build_string__ = 'v3.1.1-17-g98c44df9c_Fri Sep 21 02:39:59 2018 -0400'
 __author__ = 'Roberto Rosario'
 __author_email__ = 'roberto.rosario@mayan-edms.com'
 __description__ = 'Free Open Source Electronic Document Management System'