+++ HISTORY.rst
@@ -3,7 +3,7 @@
 * Database access in data migrations defaults to the 'default' database. Force it to the user selected database instead.
 * Don't use a hardcoded database alias for the destination of the database conversion.
 * Improve natural key support in the UserOptions model.
-
+* Update from Django 1.11.11 to 1.11.16
 
 3.1.1 (2018-09-18)
 ==================