+++ HISTORY.rst
@@ -3,8 +3,11 @@
 * Database access in data migrations defaults to the 'default' database. Force it to the user selected database instead.
 * Don't use a hardcoded database alias for the destination of the database conversion.
 * Improve natural key support in the UserOptions model.
-* Update from Django 1.11.11 to 1.11.16
+* Update from Django 1.11.11 to 1.11.16.
 * Add support to the convertdb command to operate on specified apps too.
+* Add test mixin to test the db conversion (dumping and loading) of a specific app.
+* Add an user test mixin to group user testing.
+* Add test the user managament app for database conversion.
 
 3.1.1 (2018-09-18)
 ==================